requirements.txt
rich>=13.0.0
PyYAML>=6.0.0
